add AestheticScorer class
